Last commit before sending to the employer.

+++ flutter_weather_app/lib/data/models/hourly_forecast_model.dart
@@ -8,7 +8,7 @@ class HourlyForecastModel extends HourlyForecastEntity {
     var min = json['dt_txt'].split(' ')[1].split(':')[1];
     var time = '$hour:$min';
     return HourlyForecastModel(
-      temp: json['main']['temp'],
+      temp: json['main']['temp'].toDouble(),
       time: time,
     );
   }
